Moved all controller factories under `ZF\Apigility\Admin\Controller` namespace
