+++ config/module.config.php
@@ -66,11 +66,11 @@ return [
         ],
         'factories' => [
             Controller\AppController::class            => InvokableFactory::class,
-            Controller\AuthenticationController::class => Factory\AuthenticationControllerFactory::class,
+            Controller\AuthenticationController::class => Controller\AuthenticationControllerFactory::class,
             Controller\AuthenticationType::class       => Controller\AuthenticationTypeControllerFactory::class,
-            Controller\AuthorizationController::class  => Factory\AuthorizationControllerFactory::class,
+            Controller\AuthorizationController::class  => Controller\AuthorizationControllerFactory::class,
             Controller\CacheEnabledController::class   => InvokableFactory::class,
-            Controller\ConfigController::class         => Factory\ConfigControllerFactory::class,
+            Controller\ConfigController::class         => Controller\ConfigControllerFactory::class,
             Controller\Dashboard::class                => Controller\DashboardControllerFactory::class,
             Controller\DbAutodiscovery::class          => Controller\DbAutodiscoveryControllerFactory::class,
             Controller\Documentation::class            => Controller\DocumentationControllerFactory::class,
@@ -78,14 +78,14 @@ return [
             Controller\FsPermissionsController::class  => InvokableFactory::class,
             Controller\Hydrators::class                => Controller\HydratorsControllerFactory::class,
             Controller\InputFilter::class              => Controller\InputFilterControllerFactory::class,
-            Controller\ModuleConfigController::class   => Factory\ModuleConfigControllerFactory::class,
-            Controller\ModuleCreationController::class => Factory\ModuleCreationControllerFactory::class,
+            Controller\ModuleConfigController::class   => Controller\ModuleConfigControllerFactory::class,
+            Controller\ModuleCreationController::class => Controller\ModuleCreationControllerFactory::class,
             Controller\PackageController::class        => InvokableFactory::class,
             Controller\SettingsDashboard::class        => Controller\DashboardControllerFactory::class,
-            Controller\SourceController::class         => Factory\SourceControllerFactory::class,
+            Controller\SourceController::class         => Controller\SourceControllerFactory::class,
             Controller\StrategyController::class       => InvokableFactory::class,
             Controller\Validators::class               => Controller\ValidatorsControllerFactory::class,
-            Controller\VersioningController::class     => Factory\VersioningControllerFactory::class,
+            Controller\VersioningController::class     => Controller\VersioningControllerFactory::class,
         ],
     ],
