+++ src/Controller/AuthenticationControllerFactory.php
@@ -4,37 +4,36 @@
  * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
-namespace ZF\Apigility\Admin\Factory;
+namespace ZF\Apigility\Admin\Controller;
 
 use Interop\Container\ContainerInterface;
 use Zend\ServiceManager\AbstractPluginManager;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
-use ZF\Apigility\Admin\Controller\AuthorizationController;
-use ZF\Apigility\Admin\Model\AuthorizationModelFactory;
+use ZF\Apigility\Admin\Model\AuthenticationModel;
 
-class AuthorizationControllerFactory implements FactoryInterface
+class AuthenticationControllerFactory implements FactoryInterface
 {
     /**
      * @param ContainerInterface $container
      * @param string $requestedName
      * @param null|array $options
-     * @return AuthorizationController
+     * @return AuthenticationController
      */
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
-        return new AuthorizationController($container->get(AuthorizationModelFactory::class));
+        return new AuthenticationController($container->get(AuthenticationModel::class));
     }
 
     /**
      * @param ServiceLocatorInterface $container
-     * @return AuthorizationController
+     * @return AuthenticationController
      */
     public function createService(ServiceLocatorInterface $container)
     {
         if ($container instanceof AbstractPluginManager) {
             $container = $container->getServiceLocator() ?: $container;
         }
-        return $this($container, AuthorizationController::class);
+        return $this($container, AuthenticationController::class);
     }
 }
