+++ src/Controller/AuthorizationControllerFactory.php
@@ -4,37 +4,36 @@
  * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
-namespace ZF\Apigility\Admin\Factory;
+namespace ZF\Apigility\Admin\Controller;
 
 use Interop\Container\ContainerInterface;
 use Zend\ServiceManager\AbstractPluginManager;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
-use ZF\Apigility\Admin\Controller\ModuleConfigController;
-use ZF\Configuration\ConfigResourceFactory;
+use ZF\Apigility\Admin\Model\AuthorizationModelFactory;
 
-class ModuleConfigControllerFactory implements FactoryInterface
+class AuthorizationControllerFactory implements FactoryInterface
 {
     /**
      * @param ContainerInterface $container
      * @param string $requestedName
      * @param null|array $options
-     * @return ModuleConfigController
+     * @return AuthorizationController
      */
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
-        return new ModuleConfigController($container->get(ConfigResourceFactory::class));
+        return new AuthorizationController($container->get(AuthorizationModelFactory::class));
     }
 
     /**
      * @param ServiceLocatorInterface $container
-     * @return ModuleConfigController
+     * @return AuthorizationController
      */
     public function createService(ServiceLocatorInterface $container)
     {
         if ($container instanceof AbstractPluginManager) {
             $container = $container->getServiceLocator() ?: $container;
         }
-        return $this($container, ModuleConfigController::class);
+        return $this($container, AuthorizationController::class);
     }
 }
