+++ src/Controller/ConfigControllerFactory.php
@@ -4,37 +4,36 @@
  * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
-namespace ZF\Apigility\Admin\Factory;
+namespace ZF\Apigility\Admin\Controller;
 
 use Interop\Container\ContainerInterface;
 use Zend\ServiceManager\AbstractPluginManager;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
-use ZF\Apigility\Admin\Controller\VersioningController;
-use ZF\Apigility\Admin\Model\VersioningModelFactory;
+use ZF\Configuration\ConfigResource;
 
-class VersioningControllerFactory implements FactoryInterface
+class ConfigControllerFactory implements FactoryInterface
 {
     /**
      * @param ContainerInterface $container
      * @param string $requestedName
      * @param null|array $options
-     * @return VersioningController
+     * @return ConfigController
      */
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
-        return new VersioningController($container->get(VersioningModelFactory::class));
+        return new ConfigController($container->get(ConfigResource::class));
     }
 
     /**
      * @param ServiceLocatorInterface $container
-     * @return VersioningController
+     * @return ConfigController
      */
     public function createService(ServiceLocatorInterface $container)
     {
         if ($container instanceof AbstractPluginManager) {
             $container = $container->getServiceLocator() ?: $container;
         }
-        return $this($container, VersioningController::class);
+        return $this($container, ConfigController::class);
     }
 }
