+++ src/Controller/ModuleConfigControllerFactory.php
@@ -4,37 +4,36 @@
  * @copyright Copyright (c) 2016 Zend Technologies USA Inc. (http://www.zend.com)
  */
 
-namespace ZF\Apigility\Admin\Factory;
+namespace ZF\Apigility\Admin\Controller;
 
 use Interop\Container\ContainerInterface;
 use Zend\ServiceManager\AbstractPluginManager;
 use Zend\ServiceManager\FactoryInterface;
 use Zend\ServiceManager\ServiceLocatorInterface;
-use ZF\Apigility\Admin\Controller\AuthenticationController;
-use ZF\Apigility\Admin\Model\AuthenticationModel;
+use ZF\Configuration\ConfigResourceFactory;
 
-class AuthenticationControllerFactory implements FactoryInterface
+class ModuleConfigControllerFactory implements FactoryInterface
 {
     /**
      * @param ContainerInterface $container
      * @param string $requestedName
      * @param null|array $options
-     * @return AuthenticationController
+     * @return ModuleConfigController
      */
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
-        return new AuthenticationController($container->get(AuthenticationModel::class));
+        return new ModuleConfigController($container->get(ConfigResourceFactory::class));
     }
 
     /**
      * @param ServiceLocatorInterface $container
-     * @return AuthenticationController
+     * @return ModuleConfigController
      */
     public function createService(ServiceLocatorInterface $container)
     {
         if ($container instanceof AbstractPluginManager) {
             $container = $container->getServiceLocator() ?: $container;
         }
-        return $this($container, AuthenticationController::class);
+        return $this($container, ModuleConfigController::class);
     }
 }
